+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     - Visual fixups in *ypm-menu.
     - Fixed bug in relative_path_if_subtree() that would incorrectly substitute CWD under certain circumstances.
     - Fixed various memory bugs detected with new ASAN build support.
+    - Fall back to setting C.UTF-8 locale when en_us.UTF-8 is not found and abort if we can't set a UTF-8 locale.
 ### Changed
     - Use compiler flags to force inclusion of frame pointers.
     - Changed every API call dealing with glyphs to pass by pointer since dereferencing all 4 bytes of a yed_glyph struct may be an out-of-bounds read in some cases.